requirements.txt
requests>=2.28
PyYAML>=6.0
google_play_scraper>=1.0.4   
rapidfuzz>=2.13
Pillow>=9.0
ImageHash>=4.3
python-dotenv>=1.0.0
colorama  